refactory: preview and badge
