+++ ComicSpider/spiders/jm.py
@@ -117,8 +117,8 @@ class JmSpider(BaseComicSpider2):
             self.say('') if (x + 1) % self.num_of_row == 0 else None
             frame_results[x + 1] = [url, title, preview_url]
             _likes = target.xpath('.//span[contains(@id,"albim_likes")]/text()').get()
-            likes = f"♥️{_likes.strip()}" if _likes else 0
-            preview.add(x+1, img_preview, title, preview_url, likes)
+            likes = _likes.strip() if _likes else 0
+            preview.add(x+1, img_preview, title, preview_url, likes=likes)
         self.say(preview.created_temp_html)
         self.say(font_color("<br>  jm预览图加载懂得都懂，加载不出来是正常现象哦", color='purple'))
         return self.say.frame_book_print(frame_results, url=response.url)
