+++ GUI/src/preview_format/public.js
@@ -13,4 +13,16 @@
   window.get_curr_hml = function () {
     return document.documentElement.outerHTML;
   }
+  document.addEventListener('DOMContentLoaded', function() {
+    const containers = document.querySelectorAll('div[style*="position: relative"]');
+    containers.forEach(container => {
+        const badges = container.querySelectorAll('.badge-on-img');
+        let verticalOffset = 0;
+        badges.forEach(badge => {
+            badge.style.removeProperty('top');
+            badge.style.top = `${verticalOffset}px`;
+            verticalOffset += badge.offsetHeight + 2;
+        });
+    });
+  });
 })();
